# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZIP 042 API",
    "version": "rules-2026-08-20.1 / extraction-2026-08-20",
    "description": "사회초년생을 위한 원룸 계약 안전 검사 서비스 백엔드.\n\n## 인증\n`Authorization: Bearer <Supabase access token>`\n목 모드(`ZIP042_MODE=mock`)에서는 아무 문자열이나 보내면 되고, 토큰이 곧 사용자 식별자입니다.\n\n## 시작하기\n1. `GET /v1/meta` — 기능 가용성 확인 (시세 조회 가능 여부 등)\n2. `GET /v1/addresses/search` — 주소를 골라 좌표·법정동코드를 얻습니다 (**필수**)\n3. `POST /v1/cases` — 검사 건 생성\n4. `POST /v1/cases/{caseId}/documents/upload-url` → PUT → `POST .../documents`\n5. `POST /v1/cases/{caseId}/analyze/jobs` — 비동기 분석 (권장) 또는 `POST .../analyze` (동기)\n\n## 판정의 두 축\n`verdict.score` 는 **실제 위험**만 반영합니다. 확인하지 못한 항목은 `informationGaps` 로\n따로 나가고, 그중 판정을 완료할 수 없게 만드는 것은 `blockingGaps` 입니다.\n`blockingGaps` 가 비어 있지 않으면 `contractable` 은 false 이며, 이때 UI 는\n\"위험합니다\"가 아니라 **\"아직 판단할 수 없습니다\"** 로 표현해야 합니다.\n\n## 주의\n모든 판정 응답에는 `caveats` / `disclaimer` 가 포함됩니다. **법률 자문이 아니라는 고지를**\n**UI 에 반드시 노출해야 합니다.**"
  },
  "servers": [
    {
      "url": "http://localhost:8787"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "UNPROCESSABLE",
                  "VALIDATION_FAILED",
                  "RATE_LIMITED",
                  "UPSTREAM_FAILED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "사용자에게 그대로 보여줄 수 있는 한국어 메시지"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/cases/{caseId}/special-terms": {
      "get": {
        "tags": [
          "special-terms"
        ],
        "summary": "이 집에 맞춘 추천 특약",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "추천 특약",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{caseId}/special-terms/compose": {
      "post": {
        "tags": [
          "special-terms"
        ],
        "summary": "선택한 특약을 붙여넣기용 텍스트로 조립",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "codes"
                ],
                "properties": {
                  "codes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "조립된 문구",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}